requirements.txt
Django==2.2
djangorestframework==3.9.2
envdir==1.0.1
psycopg2==2.8.2
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
elasticsearch_dsl
